refactor: adjust package structure
