+++ src/main/java/com/example/lxp/common/model/User.java
@@ -1,4 +1,4 @@
-package com.example.lxp.common.model.dto;
+package com.example.lxp.common.model;
 
 public record User(
         Long id,
